# backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QSC Platform API",
    "description": "Quantum Secure Communication platform backend. All quantum operations are classical simulations; QBER thresholds are simulation thresholds.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/communications": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Platform Communications",
        "operationId": "platform_communications_api_v1_admin_communications_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/communications": {
      "get": {
        "tags": [
          "communications"
        ],
        "summary": "List Communications",
        "operationId": "list_communications_api_v1_communications_get",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(mine|history)$",
              "default": "mine",
              "title": "Scope"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 24
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/communications/{comm_id}/timeline": {
      "get": {
        "tags": [
          "communications"
        ],
        "summary": "Communication Timeline",
        "description": "REST replay of persisted timeline events (F10 bootstrap/reconnect).",
        "operationId": "communication_timeline_api_v1_communications__comm_id__timeline_get",
        "parameters": [
          {
            "name": "comm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Comm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/communications/{comm_id}": {
      "get": {
        "tags": [
          "communications"
        ],
        "summary": "Communication Detail",
        "operationId": "communication_detail_api_v1_communications__comm_id__get",
        "parameters": [
          {
            "name": "comm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Comm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/communications/{comm_id}/recommendation": {
      "get": {
        "tags": [
          "communications"
        ],
        "summary": "Recommendation",
        "operationId": "recommendation_api_v1_communications__comm_id__recommendation_get",
        "parameters": [
          {
            "name": "comm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Comm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/communications/{comm_id}/qkd": {
      "get": {
        "tags": [
          "communications"
        ],
        "summary": "Qkd Runs",
        "operationId": "qkd_runs_api_v1_communications__comm_id__qkd_get",
        "parameters": [
          {
            "name": "comm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Comm Id"
            }
          },
          {
            "name": "include_sample",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Sample"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/communications/{comm_id}/security": {
      "get": {
        "tags": [
          "communications"
        ],
        "summary": "Security State",
        "operationId": "security_state_api_v1_communications__comm_id__security_get",
        "parameters": [
          {
            "name": "comm_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Comm Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}